Basic first pass at extracting functionality into library
